+++ src/main/java/com/nisovin/shopkeepers/shopobjects/living/types/CatShop.java
@@ -1,6 +1,5 @@
 package com.nisovin.shopkeepers.shopobjects.living.types;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.bukkit.Color;
@@ -61,9 +60,8 @@ public class CatShop extends SittableShop<Cat> {
 	}
 
 	@Override
-	public List<EditorHandler.Button> getEditorButtons() {
-		List<EditorHandler.Button> editorButtons = new ArrayList<>();
-		editorButtons.addAll(super.getEditorButtons());
+	public List<EditorHandler.Button> createEditorButtons() {
+		List<EditorHandler.Button> editorButtons = super.createEditorButtons();
 		editorButtons.add(this.getCatTypeEditorButton());
 		editorButtons.add(this.getCollarColorEditorButton());
 		return editorButtons;