+++ src/main/java/com/nisovin/shopkeepers/shopobjects/living/types/CreeperShop.java
@@ -1,6 +1,5 @@
 package com.nisovin.shopkeepers.shopobjects.living.types;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.bukkit.Material;
@@ -49,9 +48,8 @@ public class CreeperShop extends SKLivingShopObject<Creeper> {
 	}
 
 	@Override
-	public List<EditorHandler.Button> getEditorButtons() {
-		List<EditorHandler.Button> editorButtons = new ArrayList<>();
-		editorButtons.addAll(super.getEditorButtons());
+	public List<EditorHandler.Button> createEditorButtons() {
+		List<EditorHandler.Button> editorButtons = super.createEditorButtons();
 		editorButtons.add(this.getPoweredEditorButton());
 		return editorButtons;
 	}