+++ src/main/java/com/nisovin/shopkeepers/shopobjects/living/types/FoxShop.java
@@ -1,6 +1,5 @@
 package com.nisovin.shopkeepers.shopobjects.living.types;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.bukkit.Color;
@@ -64,9 +63,8 @@ public class FoxShop extends SittableShop<Fox> {
 	}
 
 	@Override
-	public List<EditorHandler.Button> getEditorButtons() {
-		List<EditorHandler.Button> editorButtons = new ArrayList<>();
-		editorButtons.addAll(super.getEditorButtons());
+	public List<EditorHandler.Button> createEditorButtons() {
+		List<EditorHandler.Button> editorButtons = super.createEditorButtons();
 		editorButtons.add(this.getFoxTypeEditorButton());
 		editorButtons.add(this.getSleepingEditorButton());
 		editorButtons.add(this.getCrouchingEditorButton());