+++ src/main/java/com/nisovin/shopkeepers/shopobjects/living/types/VillagerShop.java
@@ -1,6 +1,5 @@
 package com.nisovin.shopkeepers.shopobjects.living.types;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import org.bukkit.Color;
@@ -97,9 +96,8 @@ public class VillagerShop extends BabyableShop<Villager> {
 	}
 
 	@Override
-	public List<EditorHandler.Button> getEditorButtons() {
-		List<EditorHandler.Button> editorButtons = new ArrayList<>();
-		editorButtons.addAll(super.getEditorButtons());
+	public List<EditorHandler.Button> createEditorButtons() {
+		List<EditorHandler.Button> editorButtons = super.createEditorButtons();
 		editorButtons.add(this.getProfessionEditorButton());
 		editorButtons.add(this.getVillagerTypeEditorButton());
 		editorButtons.add(this.getVillagerLevelEditorButton());